# backend/Assets/annuity-eapp-openapi-3.yaml
openapi: 3.1.0
info:
  title: Annuity E-Application API
  description: |
    A dynamic, renderable e-application API for annuity products.

    This API is designed for two primary consumers:
    - **Frontend Applications** — Render pages, questions, and validation rules dynamically
      based on the response structure.
    - **AI Agents / LLMs** — Use `hint` fields on each question for natural language prompting,
      conditions for contextual question flow, and validation rules for inline answer validation.

    ### Key Concepts
    - **Application** — Top-level envelope containing metadata and all pages.
    - **Page** — A logical grouping of questions with optional conditional visibility.
      Pages marked with `pageRepeat` are templates rendered multiple times based on a
      source question answer (e.g., one page per surrendering company in a 1035 exchange).
    - **Question** — A single input with type, label, hint, options, visibility, and validation.
    - **Condition** — A recursive expression tree supporting simple, compound (AND/OR/NOT),
      and computed comparisons.
    - **Validation** — Rules applied at field, cross-field, group-sum, async, and allocation levels.
    - **Field Sync** — Bidirectional binding between questions that must share the same
      value (e.g., owner SSN on the main application and on each 1035 transfer form).
      Changes to any member propagate to all others; fields remain editable with a
      desync warning.

...
  version: 1.0.0
...
servers:
...
paths:
...
  /products/{productId}:
    get:
      tags:
        - Products
      summary: Get a product by productId
      operationId: getProduct
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Product object.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductSummary'
        '404':
          description: Product not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags:
        - Products
      summary: Update a product
      operationId: updateProduct
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Product updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductSummary'
        '404':
          description: Product not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - Products
      summary: Delete a product
      operationId: deleteProduct
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Product deleted.
        '404':
          description: Product not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ── Distributors ────────────────────────────────────────────
...
  /applications:
    get:
      tags:
        - Applications
      summary: List all applications
      operationId: listApplications
      responses:
        '200':
          description: Array of all application records.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ApplicationInstance'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags:
        - Applications
      summary: Create a new application
      operationId: createApplication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - productId
              properties:
                productId:
                  type: string
            examples:
              create:
                value:
                  productId: midland-fixed-annuity-001
      responses:
        '201':
          description: Application created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationInstance'
        '400':
          description: Missing productId.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Product not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /applications/{id}:
    get:
      tags:
        - Applications
      summary: Get an application by ID
      operationId: getApplication
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Application object.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationInstance'
        '404':
          description: Application not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /applications/{id}/answers:
    put:
      tags:
        - Applications
      summary: Update application answers
      operationId: updateApplicationAnswers
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - answers
              properties:
                answers:
                  type: object
            examples:
              partialUpdate:
                value:
                  answers:
                    annuitant_first_name: Jane
                    annuitant_last_name: Smith
      responses:
        '200':
          description: Updated application with merged answers.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationInstance'
        '400':
          description: Missing answers in request body.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Application not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Application already submitted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

# ─────────────────────────────────────────────────────────────
# COMPONENTS
# ─────────────────────────────────────────────────────────────
components:
  schemas:
...
    ErrorResponse:
      type: object
      description: Standard error envelope for non-validation HTTP errors.
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: Machine-readable error code for programmatic handling.
          examples:
            - PRODUCT_NOT_FOUND
        message:
          type: string
          description: Human-readable description of the error.
          examples:
            - No application definition found for product ID 'xyz-999'.
        details:
          description: Optional additional diagnostic context.
          oneOf:
            - type: object
              additionalProperties: true
            - type: 'null'
    ProductSummary:
      type: object
      properties:
        productId:
          type: string
        carrier:
          type: string
        productName:
          type: string
        productType:
          type: string
        distributors:
          type: array
          items:
            type: string
        createdAt:
          type: string
        updatedAt:
          type: string
...
    ApplicationInstance:
      type: object
      properties:
        id:
          type: string
        productId:
          type: string
        answers:
          type: object
        status:
          type: string
          enum:
            - in_progress
            - submitted
        submissionId:
          type: string
          description: Carrier submission ID, set after successful submission.
        policyNumber:
          type: string
          description: Carrier policy number, set after successful submission.
        createdAt:
          type: string
        updatedAt:
          type: string